Refactor: split Constants and ColorHistoryDrawer (Tier 2)
Finish the SwiftLint file-size cleanup so file_length can return to its
default of 400. No behaviour change.

- Split Constants.swift (442 lines) — move the PikaText localized-string
  enum into PikaText.swift (334), leaving Constants.swift at 105 (URLs,
  notification names, initial colours). Replace the file-wide
  line_length disable with two targeted disables on the only two
  genuinely-long localizer comment lines.
- Split ColorHistoryDrawer.swift (403 lines) — move ColorHistoryChip,
  PaletteNameField, and PaletteTabBar into PaletteComponents.swift (246),
  leaving the main drawer view at 159.
- Tighten .swiftlint.yml: file_length back to the default 400 now that no
  source file exceeds it.

Both targets build and the test suite passes.
